Add Ability To Call Destroy From jQuery Function
Adds the ability to call the destroy method from the jQuery function

To call the destroy method using jQuery just call the dragend method and
pass in a setting the sets the destroy property to true.  For example:

$("#dragendcont").dragend( {destroy: true } );

Without this change you need a instance of the dragend object to call
the destory method.  For example:

var instance = new Dragend();
instance.destroy();